chore: refactor and improve the keep_lease_handler
